refactoring: type별 개수 count 기능 class화 및 분리  refactoring 진행
